refactor(backend): ♻️ use trpc for memes
