+++ mongoserver/data/memes.metadata.json
@@ -6,7 +6,7 @@
       "name": "_id_"
     }
   ],
-  "uuid": "fc717f00cf3244ebb86c773c1f98b6fc",
+  "uuid": "89d5096a0b884d2b96127c7930512709",
   "collectionName": "memes",
   "type": "collection"
 }
