+++ mongoserver/data/templates.metadata.json
@@ -6,7 +6,7 @@
       "name": "_id_"
     }
   ],
-  "uuid": "3092cf9ce8b7472c8147a6606aa71473",
+  "uuid": "916e88b57dff48eeb0b76faba333ba84",
   "collectionName": "templates",
   "type": "collection"
 }
