+++ mongoserver/data/users.metadata.json
@@ -6,7 +6,7 @@
       "name": "_id_"
     }
   ],
-  "uuid": "56cfa947661946c5b2337d5973126c26",
+  "uuid": "aa54c47e9888481bb82476416137f0ab",
   "collectionName": "users",
   "type": "collection"
 }
