+++ server/index.ts
@@ -1,15 +1,17 @@
 import { createHTTPServer } from "@trpc/server/adapters/standalone";
-import { z } from "zod";
-import { publicProcedure, router } from "./src/trpc";
+
+import { router } from "./src/trpc";
 import cors from "cors";
 import { createContext } from "./src/context";
 
+import { memeRouter } from "./src/routers/meme";
 import { userRouter } from "./src/routers/user";
 import { templateRouter } from "./src/routers/template";
 
 const appRouter = router({
-  user: userRouter,
+  meme: memeRouter,
   template: templateRouter,
+  user: userRouter,
 });
 
 export type AppRouter = typeof appRouter;
