+++ server/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon index.ts",
+    "start": "nodemon index.ts",
     "prisma": "prisma generate",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
